add firebase and refactor loginVC
