+++ Navigation/GoogleService-Info.plist
@@ -1,0 +1,34 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!DOCTYPE plist PUBLIC "-//Apple//DTD PLIST 1.0//EN" "http://www.apple.com/DTDs/PropertyList-1.0.dtd">
+<plist version="1.0">
+<dict>
+	<key>CLIENT_ID</key>
+	<string>477857177257-5etrd8r6i0m2u679na7ac6lme3lrbg2m.apps.googleusercontent.com</string>
+	<key>REVERSED_CLIENT_ID</key>
+	<string>com.googleusercontent.apps.477857177257-5etrd8r6i0m2u679na7ac6lme3lrbg2m</string>
+	<key>API_KEY</key>
+	<string>AIzaSyBmzRiVTmKAF55ruFtZiw-o7LQ01V67H4E</string>
+	<key>GCM_SENDER_ID</key>
+	<string>477857177257</string>
+	<key>PLIST_VERSION</key>
+	<string>1</string>
+	<key>BUNDLE_ID</key>
+	<string>com.netology.Navigation</string>
+	<key>PROJECT_ID</key>
+	<string>navigation-22470</string>
+	<key>STORAGE_BUCKET</key>
+	<string>navigation-22470.appspot.com</string>
+	<key>IS_ADS_ENABLED</key>
+	<false></false>
+	<key>IS_ANALYTICS_ENABLED</key>
+	<false></false>
+	<key>IS_APPINVITE_ENABLED</key>
+	<true></true>
+	<key>IS_GCM_ENABLED</key>
+	<true></true>
+	<key>IS_SIGNIN_ENABLED</key>
+	<true></true>
+	<key>GOOGLE_APP_ID</key>
+	<string>1:477857177257:ios:053da96c0a414490d213b8</string>
+</dict>
+</plist>
